backend/requirements.txt
telethon>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0

# dev/test
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
